requirements.txt
streamlit
urllib3
python-dotenv